refactor: 'paths' configuration is removed in tsconfig due to ts-node-dev incompatibility
