+++ README.md
@@ -1,0 +1,3 @@
+'''
+pnpm prisma:generate
+'''
