+++ src/utils/encryption.ts
@@ -1,4 +1,4 @@
-import bcrypt from 'bcryptjs';
+import bcrypt from 'bcrypt';
 
 export const encryptPassword = async (password: string): Promise<string> => {
   const hashedPassword = await bcrypt.hash(password, 10);
