+++ tsconfig.json
@@ -29,7 +29,7 @@
     "rootDir": "./src",                                  /* Specify the root folder within your source files. */
     "moduleResolution": "nodenext",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
-    "paths": { "@/*": ["./src/*"] },                     /* Specify a set of entries that re-map imports to additional lookup locations. */
+    // "paths": { "@/*": ["./src/*"] },                     /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
     // "types": [],                                      /* Specify type package names to be included without being referenced in a source file. */
